requirements.txt
alpaca-py==0.43.2
pandas==2.3.3
numpy==2.4.4
python-dotenv==1.2.2
yfinance==1.3.0
